Connection整合 flask_test.py 調整部分static connection呼叫 將Config/ApiTestPC(部分)function移至 Connection.py 調整部分報錯寫法
